refactor(build, core): Migrate to PEP 420 Namespace Packages
This commit refactors the project to use a pure PEP 420 implicit namespace package structure, replacing the legacy pkgutil-based approach.

Key changes include:
- Switched the build system in pyproject.toml from hatchling to setuptools to properly support PEP 420 namespaces.
- Removed __init__.py from namespace directories (src/rhosocial/ and src/rhosocial/activerecord/) to enable namespace package discovery.
- Moved the core ActiveRecord class from the __init__.py file to a new dedicated module: src/rhosocial/activerecord/model.py.
- Deleted the legacy setup.py file as all build configuration is now in pyproject.toml.

This migration improves maintainability, aligns with modern Python packaging standards, and resolves module resolution issues in IDEs and static analysis tools.
